use a singleton for the modifier manager
